Expose keras training loss metric of singular output
Previously, the loss metric of the individual output was not include in the logging metrics when the model only had a singular output.
It seems this was done under the assumption that with only a single output, that the overall model loss would be the same, and so the metric would be redundant. However model loss is comprised of other sources, such as regularization losses, and as such these metrics are not guaranteed to be the same.
Having visibility into the output loss, independent of other losses in the model can provide better visibility and insight into the model during training.